Fixed evaluation of existing backup files and added check for free space
Fixed: Evaluation of existing backup files was failing because the size
check was occurring within the if statement, where it was needed in the
elseif and else clauses. Moved the filesize check prior to the if so
that it is available throughout.
Added: function to check for sufficient free space on offload/backup
location before copying a backup file.