Clear matches array between base and incr file search
If the base file is found but the incremental file is not, the script was previously keeping the base file name in the $matches variable. This would fool the check for incremental file existence, because the test merely checks for matches from the previous query, but the query does not return an empty array if no matches are found, it just leaves the array alone.

Resolved this by clearing the $matches variable between tests.